requirements.txt
# requirements.txt
Flask==2.3.2
Werkzeug==2.3.4
urllib3==1.26.3
python-dotenv==1.0.0
django-environ==0.10.0
easyocr
torch
torchvision>=0.5
opencv-python-headless
scipy
numpy
Pillow
scikit-image
python-bidi
PyYAML
Shapely
pyclipper
ninja
